Feat: Add a cap to locally stored log file count

+++ samples/IoT-Edge/Python3/Diagnosis/storageClient.py
@@ -19,13 +19,24 @@ class storageClient():
         return self.azure_blob_storage_container
 
     def upload_log_to_storage(self, filepath):
-        print(time.strftime('%Y/%m/%d %H:%M:%S') + ' [Upload Log] Upload start. Container = ' + str(self.azure_blob_storage_container) + ', File = ' + filepath)
+        print(time.strftime('%Y/%m/%d %H:%M:%S') + ' [Upload Log] Start. Container = ' + str(self.azure_blob_storage_container) + ', File = ' + filepath)
         try:
             blob_client = self.blob_service_client.get_blob_client(container=self.azure_blob_storage_container, blob=os.path.basename(filepath))
         except:
-            print(time.strftime('%Y/%m/%d %H:%M:%S') + ' Error: Blob or container does not exist')
-            return None
+            print(time.strftime('%Y/%m/%d %H:%M:%S') + ' [Upload Log] Blob storage or container does not exist.')
+            return 1, 'Failed to connect to the target blob storage, please check the blob connection string and container name settings.'
+        if not blob_client:
+            print(time.strftime('%Y/%m/%d %H:%M:%S') + ' [Upload Log] Failed to connect to the target container.')
+            return 1, 'Failed to connect to the target blob storage, please check the blob connection string and container name settings.'
+
         content_settings = ContentSettings(content_type='application/zip')
-        blob_client.upload_blob(open(filepath, 'rb'), overwrite=True, content_settings=content_settings)
-        os.remove(filepath)
-        return blob_client.url
+        try:
+            blob_client.upload_blob(open(filepath, 'rb'), overwrite=True, content_settings=content_settings, timeout=10)
+            print(time.strftime('%Y/%m/%d %H:%M:%S') + ' [Upload Log] Complete, URL = ' + blob_client.url)
+            os.remove(filepath)
+            return 0, blob_client.url
+        except Exception as e:
+            print(time.strftime('%Y/%m/%d %H:%M:%S') + ' [Upload Log] Failed.')
+            print(e)
+            return 2, 'Failed to upload log, please check the blob settings and network connectivity.'
+
